+++ app/contributor-preferences.js
@@ -1,0 +1,46 @@
+const { ipcRenderer } = require('electron')
+const HtmlTranslate = require('./utils/htmlTranslate')
+
+const htmlTranslate = new HtmlTranslate(document)
+let eventsAttached = false
+
+window.onload = (event) => {
+  ipcRenderer.send('send-settings')
+  htmlTranslate.translate()
+  setTimeout(() => { eventsAttached = true }, 500)
+}
+
+document.ondragover = event =>
+  event.preventDefault()
+
+document.ondrop = event =>
+  event.preventDefault()
+
+ipcRenderer.on('renderSettings', (event, settings) => {
+  document.querySelectorAll('input[type="checkbox"]').forEach(checkbox => {
+    const isNegative = checkbox.classList.contains('negative')
+    checkbox.checked = isNegative ? !settings[checkbox.value] : settings[checkbox.value]
+    if (!eventsAttached) {
+      checkbox.onchange = (event) =>
+        ipcRenderer.send('save-setting', checkbox.value,
+          isNegative ? !checkbox.checked : checkbox.checked)
+    }
+  })
+
+  document.querySelectorAll('input[type="range"]').forEach(range => {
+    const divisor = range.dataset.divisor
+    range.value = settings[range.name] / divisor
+    range.closest('div').querySelector('output').innerHTML = range.value
+    if (!eventsAttached) {
+      range.onchange = event => {
+        range.closest('div').querySelector('output').innerHTML = range.value
+        ipcRenderer.send('save-setting', range.name, range.value * divisor)
+      }
+      range.oninput = event => {
+        range.closest('div').querySelector('output').innerHTML = range.value
+      }
+    }
+  })
+
+  htmlTranslate.translate()
+})